extract node edge simplification logic to lib module
